+++ packages/next-admin/src/utils/server.ts
@@ -1,22 +1,23 @@
 import { Prisma, PrismaClient } from "@prisma/client";
+import formidable from "formidable";
+import { IncomingMessage } from "http";
+import { Writable } from "stream";
 import {
+  AdminFormData,
   EditFieldsOptions,
   EditOptions,
   Enumeration,
   Field,
-  AdminFormData,
   ListOptions,
   ModelName,
   ModelWithoutRelationships,
   NextAdminOptions,
+  ObjectField,
   ScalarField,
-  Schema,
+  Schema
 } from "../types";
 import { createWherePredicate } from "./prisma";
 import { isNativeFunction, uncapitalize } from "./tools";
-import { IncomingMessage } from "http";
-import formidable from "formidable";
-import { Writable } from "stream";
 
 export const models = Prisma.dmmf.datamodel.models;
 export const resources = models.map((model) => model.name as ModelName);
@@ -36,11 +37,52 @@ export const getResources = (
   options?: NextAdminOptions
 ): Prisma.ModelName[] => {
   const definedModels =
-    options && options.model
+    options?.model
       ? (Object.keys(options.model) as Prisma.ModelName[])
       : [];
   return definedModels.length > 0 ? definedModels : resources;
 };
+
+export const getToStringForRelations = <M extends ModelName>(modelName: M, fieldName: Field<M>, modelNameRelation: ModelName, options?: NextAdminOptions, relationToFields?: any[]) => {
+  const nonChekedToString =
+    // @ts-expect-error
+    options?.model?.[modelName]?.edit?.fields?.[fieldName]?.optionFormatter
+    || options?.model?.[modelNameRelation]?.toString;
+  const modelRelationIdField = getModelIdProperty(modelNameRelation);
+  const toStringForRelations =
+    (nonChekedToString && !isNativeFunction(nonChekedToString))
+      ? nonChekedToString
+      : (item: any) =>
+        item[relationToFields?.[0]] ?? item[modelRelationIdField];
+
+  return toStringForRelations;
+}
+
+/**
+ * Order the fields in the schema according to the display option
+ * 
+ * @param schema
+ * @param resource
+ * @param options
+ * 
+ * @returns schema
+ */
+export const orderSchema = (schema: Schema, resource: ModelName, options?: NextAdminOptions) => {
+  const modelName = resource;
+  const model = models.find((model) => model.name === modelName);
+  if (!model) return schema;
+  const edit = options?.model?.[modelName]?.edit as EditOptions<typeof modelName>;
+  const display = edit?.display;
+  if (display) {
+    const properties = schema.definitions[modelName].properties;
+    const propertiesOrdered = {} as Record<string, any>;
+    display.forEach((property) => {
+      propertiesOrdered[property] = properties[property];
+    });
+    schema.definitions[modelName].properties = propertiesOrdered;
+  }
+  return schema;
+}
 
 /**
  * Fill fields with relations with the values of the related model, and inject them into the schema
@@ -49,33 +91,55 @@ export const getResources = (
  * @param prisma
  * @param requestOptions
  * @param options
+ * 
+ * @returns schema
  */
 export const fillRelationInSchema = async (
   schema: Schema,
   prisma: PrismaClient,
   resource: ModelName,
   requestOptions: any,
-  options?: NextAdminOptions
+  options?: NextAdminOptions,
 ) => {
   const modelName = resource;
   const model = models.find((model) => model.name === modelName);
-  if (!model) return schema;
+  const display = options?.model?.[modelName]?.edit?.display;
+  let fields;
+  if (model?.fields && display) {
+    // @ts-expect-error
+    fields = model.fields.filter((field) => display.includes(field.name));
+  } else {
+    fields = model?.fields;
+  }
+
+  if (!model || !fields) return schema;
   await Promise.all(
-    model.fields.map(async (field) => {
+    fields.map(async (field) => {
       const fieldName = field.name as Field<typeof modelName>;
       const fieldType = field.type;
       const fieldKind = field.kind;
-      const relationFromFields = field.relationFromFields;
       const relationToFields = field.relationToFields;
+      const relationFromFields = field.relationFromFields;
+
       if (fieldKind === "enum") {
         const fieldValue =
           schema.definitions[modelName].properties[
-            field.name as Field<typeof modelName>
+          field.name as Field<typeof modelName>
           ];
         if (fieldValue) {
           fieldValue.enum = fieldValue.enum?.map((item) =>
             typeof item !== "object" ? { label: item, value: item } : item
           );
+          const search = requestOptions[`${fieldName}search`];
+          if (search) {
+            fieldValue.enum = fieldValue.enum?.filter((item: any) =>
+              item.label.toLowerCase().includes(search.toLowerCase())
+            );
+          }
+        }
+
+        if (fieldValue?.default) {
+          fieldValue.default = typeof fieldValue.default !== "object" ? { label: fieldValue.default, value: fieldValue.default } : fieldValue.default;
         }
       }
       if (fieldKind === "object") {
@@ -85,25 +149,15 @@ export const fillRelationInSchema = async (
         );
         const listOptions = options?.model?.[modelNameRelation]
           ?.list as ListOptions<typeof modelNameRelation>;
-        const optionsForRelations =
-          listOptions?.search ?? remoteModel?.fields.map((field) => field.name);
-        const relationProperty: Field<typeof modelName> =
-          (relationFromFields?.[0] as Field<typeof modelName>) ?? fieldName;
+        const optionsForRelations = listOptions?.search ?? remoteModel?.fields.map((field) => field.name);
         const fieldsFiltered = remoteModel?.fields.filter(
           (field) => (optionsForRelations as string[])?.includes(field.name)
         );
-        const search = requestOptions[`${relationProperty}search`];
-        const where = createWherePredicate(fieldsFiltered, search);
-        const nonChekedToString = options?.model?.[modelNameRelation]?.toString;
         const modelRelationIdField = getModelIdProperty(modelNameRelation);
-        const toStringForRelations =
-          nonChekedToString && !isNativeFunction(nonChekedToString)
-            ? nonChekedToString
-            : (item: any) =>
-                item[relationToFields?.[0]] ?? item[modelRelationIdField];
+        const search = requestOptions[`${fieldName}search`];
+        const where = createWherePredicate(fieldsFiltered, search);
+        const toStringForRelations = getToStringForRelations(modelName, fieldName, modelNameRelation, options, relationToFields);
         if (
-          relationFromFields &&
-          relationFromFields.length > 0 &&
           relationToFields!.length > 0
         ) {
           //Relation One-to-Many, Many side
@@ -113,6 +167,7 @@ export const fillRelationInSchema = async (
             // @ts-expect-error
             .findMany({
               where,
+              take: 20,
             })
             .then((data: any[]) =>
               data.forEach((item) => {
@@ -122,15 +177,25 @@ export const fillRelationInSchema = async (
                 });
               })
             );
-          schema.definitions[modelName].properties[relationProperty] = {
+          schema.definitions[modelName].properties[fieldName] = {
             type: "string",
             enum: enumeration,
           };
-          delete schema.definitions[modelName].properties[fieldName];
+
+          const required = schema.definitions[modelName].required;
+          const relationFromFieldsRequired = relationFromFields?.every(
+            (field) => required?.includes(field)
+          );
+
+          if (relationFromFieldsRequired) {
+            required?.push(fieldName);
+            schema.definitions[modelName].required = required;
+          }
+
         } else {
           const fieldValue =
             schema.definitions[modelName].properties[
-              field.name as Field<typeof modelName>
+            field.name as Field<typeof modelName>
             ];
           if (fieldValue) {
             let enumeration: Enumeration[] = [];
@@ -138,6 +203,7 @@ export const fillRelationInSchema = async (
               // @ts-expect-error
               .findMany({
                 where,
+                take: 20,
               })
               .then((data: any[]) =>
                 data.forEach((item) => {
@@ -174,7 +240,8 @@ export const fillRelationInSchema = async (
 export const transformData = <M extends ModelName>(
   data: any,
   resource: M,
-  editOptions: EditOptions<M>
+  editOptions: EditOptions<M>,
+  options: NextAdminOptions
 ) => {
   const modelName = resource;
   const model = models.find((model) => model.name === modelName);
@@ -186,27 +253,25 @@ export const transformData = <M extends ModelName>(
     const get = editOptions?.fields?.[key as Field<M>]?.handler?.get;
     if (get) {
       acc[key] = get(data[key]);
-    } else {
-      if (fieldKind === "object") {
-        const modelRelation = field!.type as ModelName;
-
-        const modelRelationIdField = getModelIdProperty(modelRelation);
-
-        // Flat relationships to id
-        if (Array.isArray(data[key])) {
-          acc[key] = data[key].map((item: any) => item[modelRelationIdField]);
-        } else {
-          acc[key] = data[key] ? data[key][modelRelationIdField] : null;
-        }
+    } else if (fieldKind === "enum") {
+      acc[key] = data[key] ? { label: data[key], value: data[key] } : null;
+    } else if (fieldKind === "object") {
+      const modelRelation = field!.type as ModelName;
+      const modelRelationIdField = getModelIdProperty(modelRelation);
+      const toStringForRelations = getToStringForRelations(modelName, key as Field<M>, modelRelation, options);
+      if (Array.isArray(data[key])) {
+        acc[key] = data[key].map((item: any) => ({ label: toStringForRelations(item), value: item[modelRelationIdField] }));
       } else {
-        const fieldTypes = field?.type;
-        if (fieldTypes === "DateTime") {
-          acc[key] = data[key] ? data[key].toISOString() : null;
-        } else if (fieldTypes === "Json") {
-          acc[key] = data[key] ? JSON.stringify(data[key]) : null;
-        } else {
-          acc[key] = data[key] ? data[key] : null;
-        }
+        acc[key] = data[key] ? { label: toStringForRelations(data[key]), value: data[key][modelRelationIdField] } : null;
+      }
+    } else {
+      const fieldTypes = field?.type;
+      if (fieldTypes === "DateTime") {
+        acc[key] = data[key] ? data[key].toISOString() : null;
+      } else if (fieldTypes === "Json") {
+        acc[key] = data[key] ? JSON.stringify(data[key]) : null;
+      } else {
+        acc[key] = data[key] ? data[key] : null;
       }
     }
     return acc;
@@ -237,15 +302,14 @@ export const findRelationInData = async (
         dmmfPropertyRelationFromFields!.length > 0 &&
         dmmfPropertyRelationToFields!.length > 0
       ) {
+        const idProperty = getModelIdProperty(dmmfProperty.type as ModelName)
         data.map((item) => {
           if (item[dmmfPropertyName]) {
             item[dmmfPropertyName] = {
               type: "link",
               value: {
                 label: item[dmmfPropertyName],
-                url: `${dmmfProperty.type as ModelName}/${
-                  item[dmmfPropertyName]["id"]
-                }`,
+                url: `${dmmfProperty.type as ModelName}/${item[dmmfPropertyName][idProperty]}`,
               },
             };
           } else {
@@ -317,6 +381,13 @@ export const parseFormData = <M extends ModelName>(
   return parsedData;
 };
 
+export const formatId = (ressource: ModelName, id: string) => {
+  const model = models.find((model) => model.name === ressource);
+  const idProperty = getModelIdProperty(ressource);
+  return model?.fields.find((field) => field.name === idProperty)?.type === "Int"
+    ? Number(id)
+    : id;
+}
 /**
  * Convert the form data to the format expected by Prisma
  *
@@ -337,20 +408,15 @@ export const formattedFormData = async <M extends ModelName>(
   await Promise.allSettled(
     dmmfSchema.map(async (dmmfProperty) => {
       if (dmmfProperty.name in formData) {
-        const dmmfPropertyName = dmmfProperty.name as Field<M>;
         const dmmfPropertyType = dmmfProperty.type;
         const dmmfPropertyKind = dmmfProperty.kind;
         if (dmmfPropertyKind === "object") {
+          const dmmfPropertyName = dmmfProperty.name as keyof ObjectField<M>;
           const dmmfPropertyTypeTyped = dmmfPropertyType as Prisma.ModelName;
           const fieldValue =
             schema.definitions[modelName].properties[
-              dmmfPropertyName as Field<typeof dmmfPropertyTypeTyped>
+            dmmfPropertyName as Field<typeof dmmfPropertyTypeTyped>
             ];
-          const model = models.find((model) => model.name === dmmfPropertyType);
-          const formatId = (value?: string) =>
-            model?.fields.find((field) => field.name === "id")?.type === "Int"
-              ? Number(value)
-              : value;
           if (fieldValue?.type === "array") {
             formData[dmmfPropertyName] = JSON.parse(
               formData[dmmfPropertyName]!
@@ -358,17 +424,17 @@ export const formattedFormData = async <M extends ModelName>(
             formattedData[dmmfPropertyName] = {
               // @ts-expect-error
               [creating ? "connect" : "set"]: formData[dmmfPropertyName].map(
-                (item: any) => ({ id: formatId(item) })
+                (item: any) => ({ id: formatId(dmmfPropertyType as ModelName, item) })
               ),
             };
           } else {
             const connect = Boolean(formData[dmmfPropertyName]);
-            if (!creating)
-              formattedData[dmmfPropertyName] = connect
-                ? { connect: { id: formatId(formData[dmmfPropertyName]) } }
-                : { disconnect: true };
+            formattedData[dmmfPropertyName] = connect
+              ? { connect: { id: formatId(dmmfPropertyType as ModelName, formData[dmmfPropertyName]!) } }
+              : !creating && { disconnect: true }
           }
         } else {
+          const dmmfPropertyName = dmmfProperty.name as keyof ScalarField<M>;
           if (dmmfPropertyType === "Int") {
             formattedData[dmmfPropertyName] = !isNaN(
               Number(formData[dmmfPropertyName])
@@ -410,7 +476,7 @@ export const formattedFormData = async <M extends ModelName>(
               if (typeof uploadResult !== "string") {
                 console.warn(
                   "Upload handler must return a string, fallback to no-op for field " +
-                    dmmfPropertyName.toString()
+                  dmmfPropertyName.toString()
                 );
               } else {
                 formattedData[dmmfPropertyName] = uploadResult;
@@ -423,7 +489,6 @@ export const formattedFormData = async <M extends ModelName>(
       }
     })
   );
-
   return formattedData;
 };
 
@@ -469,7 +534,7 @@ export const changeFormatInSchema = <M extends ModelName>(
     const dmmfPropertyName = dmmfProperty.name as Field<typeof modelName>;
     const fieldValue =
       schema.definitions[modelName].properties[
-        dmmfPropertyName as Field<typeof modelName>
+      dmmfPropertyName as Field<typeof modelName>
       ];
 
     if (fieldValue && dmmfProperty.type === "Json") {
@@ -496,7 +561,7 @@ export const removeHiddenProperties = <M extends ModelName>(
   resource: M,
   editOptions: EditOptions<M>
 ) => {
-  if (!editOptions) return schema;
+  if (!editOptions?.display) return schema;
   const properties = schema.definitions[resource].properties;
   Object.keys(properties).forEach((property) => {
     if (!editOptions.display?.includes(property as Field<M>)) {
@@ -571,18 +636,7 @@ export const getResourceIdFromParam = (param: string, resource: ModelName) => {
   if (param === "new") {
     return undefined;
   }
-
-  const model = models.find((model) => model.name === resource);
-
-  const idType = model?.fields.find(
-    (field) => field.name === getModelIdProperty(resource)
-  )?.type;
-
-  if (idType === "Int") {
-    return Number(param);
-  }
-
-  return param;
+  return formatId(resource, param);
 };
 
 export const getFormDataValues = async (req: IncomingMessage) => {
@@ -647,7 +701,6 @@ export const getFormValuesFromFormData = async (formData: FormData) => {
     if (key.startsWith("$ACTION")) {
       return;
     }
-
     tmpFormValues[key] = val;
   });